Interlocking_Diagram/requirements.txt
detecto==1.1.4
Flask==1.1.2
Pillow==7.2.0
